--- packages/shared-base/src/Routes/PopupModalRoutes.ts
@@ -1,7 +1,0 @@
-export enum PopupModalRoutes {
-    ChooseNetwork = '/modal/choose-network',
-    SwitchWallet = '/modal/switch-wallet',
-    ConnectSocialAccount = '/modal/connect-social-account',
-    SelectProvider = '/modal/select-provider',
-    ConnectProvider = '/modal/connect-provider',
-}
