+++ packages/shared-base/src/helpers/index.ts
@@ -16,3 +16,4 @@ export * from './parseURL.js'
 export * from './purify.js'
 export * from './pvtutils.js'
 export * from './safeEmptyList.js'
+export * from './relativeRouteOf.js'
