+++ packages/shared-base/src/index.ts
@@ -11,17 +11,15 @@ export * from './Messages/index.js'
 export * from './NextID/index.js'
 export * from './NextID/types.js'
 export * from './Pageable/index.js'
-export * from './Routes/index.js'
 export * from './Site/index.js'
 export * from './Site/types.js'
 export * from './Sniffings/index.js'
 export * from './Subscriptions/index.js'
-export * from './collections/index.js'
 export * from './i18n/index.js'
 export * from './kv-storage/index.js'
 export * from './onDemandWorker/index.js'
 export * from './SingletonModal/index.js'
 export * from './serializer/index.js'
-export * from './legacy-settings/index.js'
+export * from './LegacySettings/index.js'
 export * from './settings/index.js'
 export * from './Telemetry/index.js'
