+++ packages/shared-base/src/types/index.ts
@@ -7,3 +7,4 @@ export * from './MIME.js'
 export * from './NameService.js'
 export * from './Wallet.js'
 export * from './SocialIdentity.js'
+export * from './Observable.js'
